Rework the clang-tidy config and fix the defects it found (#2482)
* Rewrite the clang-tidy config as an allow-list

`Checks: *` meant every LLVM upgrade added checks unasked, and the noise
buried the findings: on LLVM 20 the top three by volume were
misc-include-cleaner (532), performance-enum-size (466) and
special-member-functions (311), none of them defects. Enable families
deliberately instead.

Scope is magda/ only. Tests are out: bugprone-chained-comparison fired 311
times there, every one of them Catch2's REQUIRE(a == b) expanding to
`Decomposer() <= a == b`.

HeaderFilterRegex was unanchored, so `tests` matched
third_party/JUCE/.../unit_tests/*.h while magda/engine was never covered at
all. Anchor it and exclude third_party explicitly.

The lint targets only ever walked magda/daw, silently skipping engine,
agents and scripting.

Drop the note claiming performance-noexcept-move-constructor and
bugprone-pointer-arithmetic-on-polymorphic-object crash the run: a sweep of
all 885 magda/ TUs with both enabled completed with no crash and no findings.


* Fix use-after-move, integer division and rounding defects

Three use-after-move bugs, all of which changed behaviour:

- TranscriptionService moved onComplete into the `fail` lambda's capture,
  then moved it again into the shared Completion. Every success path got an
  empty std::function, so transcription completion never fired. The shared
  Completion is now built first and `fail` draws on it.

- AIChatConsoleContent moved `error` into ConsoleRunOutput, then tested
  error.empty() 53 lines later to decide whether to append the mixing-agent
  caveat. A moved-from string reads empty, so the caveat appeared even when
  the run had errored. Captured as `hadError` before the move.

- ChordPanelContent moved paintData into the container, then tested
  paintData.rows.empty() for the loading-height branch - always true after
  the move, so the prompt height was applied whatever the row count. Reads
  through the container now.

Seven integer divisions feeding float positions truncated sub-pixel offsets
in note, marker, tick and pill centring. Four `static_cast<int>(x + 0.5)`
roundings are now juce::roundToInt, which is also correct for negatives.


* Drop redundant optional unwrapping and mark deliberate release() discards

Thirteen sites unwrapped a std::optional only for the result to be wrapped
straight back into one - `stop(*due)` into stop(std::optional<double>),
`f.bpm = *p` into an optional field, and the
`targetBeat ? *targetBeat : std::optional<te::MonotonicBeat>{}` spelling of
`targetBeat`. Each sits behind a guard that already proved the optional
engaged, so passing it through is equivalent and drops a dereference that
would be undefined if a guard were ever moved.

The two discarded unique_ptr::release() calls hand ownership to a JUCE
AudioFormatWriter and are meant to drop the pointer; static_cast<void> says
so rather than leaving it looking like an oversight.


* Remove branches whose then and else did the same thing

NodeComponent looked up the device three times in the mod-matrix callbacks
and then called the same TrackManager method whichever way the branch went.
The lookup fed nothing; onModLinkEnabledChanged alongside it already does the
same work without one.

TimelineController ran the same two assignments on both edges of the playing
flag, and CurvePointComponent called onPointSelected(pointId_) identically in
both arms of a cmd/shift test - the callback carries no modifier, so a
modified click cannot mean anything different. That arm is now one call with
a note saying the extend-selection case is not wired up, rather than a branch
that looks like it handles it.


* Size string_view conversions, add menu default cases, mark the rest NOLINT

juce::String(sv.data()) reads to a NUL that a string_view does not promise.
The four API_VERSION conversions and the tonal-band label survive only
because the views happen to name literals; they now pass .size() so a view
built from a substring cannot walk off the end.

Three menu-result switches gained `default: break;` for the dismissed case,
which they already relied on.

The remaining bugprone findings are wrong here and are suppressed one site at
a time, with the reason, rather than by disabling the checks - they are worth
keeping for new code. signed-char-misuse would have been actively harmful:
AudioFeatures and PlanExecutor need the sign to carry a no-pitch-class
sentinel and a downward transpose, and casting to unsigned char first would
have broken both. The two log macros use NOLINTBEGIN/END because a
NOLINTNEXTLINE inside a backslash-continued #define ends the continuation.


* Drop the repeated NOLINTs into the config, keep the rounding as it was

Nineteen files were touched only to add a NOLINT comment. The five checks
behind them are off in .clang-tidy instead, with the same reasons recorded
there once rather than at each site. Three genuine one-offs stay inline: an
MSVC macro name that has to be spelled exactly, an operator= whose
self-assignment reloads and stores the same value, and a selector that returns
one of its reference parameters.

The four ClipWaveformPainter roundings go back to static_cast<int>(x + 0.5).
juce::roundToInt rounds half to even, the old form rounds half up, and these
are pixel positions - a segment edge could land a pixel over. Nothing here can
be negative, which is the case the check exists for, so the change was a
rendering difference for no defect.

+++ magda/daw/ui/panels/content/DrumGridClipContent.cpp
@@ -1861,7 +1861,7 @@ class DrumGridRowLabels : public juce::Component {
                         const int pillW = juce::jmax(18, shortTag.length() * 8 + 8);
                         juce::Rectangle<float> pill(
                             static_cast<float>(textX),
-                            static_cast<float>(y + (rowHeight_ - pillH) / 2),
+                            static_cast<float>(y) + static_cast<float>(rowHeight_ - pillH) / 2.0f,
                             static_cast<float>(pillW), static_cast<float>(pillH));
                         g.setColour(
                             DarkTheme::getColour(DarkTheme::ACCENT_PRIMARY).withAlpha(0.85f));
